requirements.txt
yara-python
requests
prettytable
colorama